refactor(ChildImage): refactor child image package

+++ src/main/java/com/cognitivecare4u/cognitiveapi/visual_cognition/ClassifierClass.java
@@ -11,5 +11,5 @@ public class ClassifierClass {
 
     private String name;
 
-    private double score;
+    private float score;
 }
